Finished Day 1 Part 1
This is my first time reading from a file line by line in Rust. I did
not want to just load everything into a string somewhere on the Heap, so
I practiced working with BufReader. The Yandex Course is actually
helpfull. I also worked with iterators, and learned about the `next()`
method.

Initially I made the bug of doing `.nth(1)` and `.nth(2)`, but it is
wrong, because I am advancing the position. And I should have used
`.nth(1)` followed by `.nth(0)` But I instead choose to skip, and then
use the next next